+++ src/main/java/org/mybatis/dynamic/sql/select/QueryExpressionDSL.java
@@ -24,6 +24,7 @@ import java.util.function.Consumer;
 
 import org.jetbrains.annotations.NotNull;
 import org.mybatis.dynamic.sql.BasicColumn;
+import org.mybatis.dynamic.sql.BindableColumn;
 import org.mybatis.dynamic.sql.CriteriaGroup;
 import org.mybatis.dynamic.sql.SortSpecification;
 import org.mybatis.dynamic.sql.SqlTable;
@@ -349,12 +350,12 @@ public class QueryExpressionDSL<R>
             this.joinType = joinType;
         }
 
-        public JoinSpecificationFinisher on(BasicColumn joinColumn, JoinCondition joinCondition) {
+        public <T> JoinSpecificationFinisher on(BindableColumn<T> joinColumn, JoinCondition<T> joinCondition) {
             return new JoinSpecificationFinisher(joinTable, joinColumn, joinCondition, joinType);
         }
 
-        public JoinSpecificationFinisher on(BasicColumn joinColumn, JoinCondition onJoinCondition,
-                JoinCriterion... andJoinCriteria) {
+        public <T> JoinSpecificationFinisher on(BindableColumn<T> joinColumn, JoinCondition<T> onJoinCondition,
+                JoinCriterion<?>... andJoinCriteria) {
             return new JoinSpecificationFinisher(joinTable, joinColumn, onJoinCondition, joinType, andJoinCriteria);
         }
     }
@@ -364,9 +365,9 @@ public class QueryExpressionDSL<R>
             implements Buildable<R> {
         private final JoinSpecification.Builder joinSpecificationBuilder;
 
-        public JoinSpecificationFinisher(TableExpression table, BasicColumn joinColumn,
-                JoinCondition joinCondition, JoinType joinType) {
-            JoinCriterion joinCriterion = new JoinCriterion.Builder()
+        public <T> JoinSpecificationFinisher(TableExpression table, BindableColumn<T> joinColumn,
+                JoinCondition<T> joinCondition, JoinType joinType) {
+            JoinCriterion<T> joinCriterion = new JoinCriterion.Builder<T>()
                     .withConnector("on") //$NON-NLS-1$
                     .withJoinColumn(joinColumn)
                     .withJoinCondition(joinCondition)
@@ -379,9 +380,9 @@ public class QueryExpressionDSL<R>
             addJoinSpecificationBuilder(joinSpecificationBuilder);
         }
 
-        public JoinSpecificationFinisher(TableExpression table, BasicColumn joinColumn,
-                JoinCondition joinCondition, JoinType joinType, JoinCriterion... andJoinCriteria) {
-            JoinCriterion onJoinCriterion = new JoinCriterion.Builder()
+        public <T> JoinSpecificationFinisher(TableExpression table, BindableColumn<T> joinColumn,
+                JoinCondition<T> joinCondition, JoinType joinType, JoinCriterion<?>... andJoinCriteria) {
+            JoinCriterion<T> onJoinCriterion = new JoinCriterion.Builder<T>()
                     .withConnector("on") //$NON-NLS-1$
                     .withJoinColumn(joinColumn)
                     .withJoinCondition(joinCondition)
@@ -412,8 +413,8 @@ public class QueryExpressionDSL<R>
             return QueryExpressionDSL.this.where();
         }
 
-        public JoinSpecificationFinisher and(BasicColumn joinColumn, JoinCondition joinCondition) {
-            JoinCriterion joinCriterion = new JoinCriterion.Builder()
+        public <T> JoinSpecificationFinisher and(BindableColumn<T> joinColumn, JoinCondition<T> joinCondition) {
+            JoinCriterion<T> joinCriterion = new JoinCriterion.Builder<T>()
                     .withConnector("and") //$NON-NLS-1$
                     .withJoinColumn(joinColumn)
                     .withJoinCondition(joinCondition)