+++ src/main/java/org/mybatis/dynamic/sql/select/render/JoinRenderer.java
@@ -18,9 +18,11 @@ package org.mybatis.dynamic.sql.select.render;
 import static org.mybatis.dynamic.sql.util.StringUtilities.spaceBefore;
 
 import java.util.Objects;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
 import org.mybatis.dynamic.sql.BasicColumn;
+import org.mybatis.dynamic.sql.render.RenderingStrategy;
 import org.mybatis.dynamic.sql.render.TableAliasCalculator;
 import org.mybatis.dynamic.sql.select.join.JoinCriterion;
 import org.mybatis.dynamic.sql.select.join.JoinModel;
@@ -32,11 +34,15 @@ public class JoinRenderer {
     private final JoinModel joinModel;
     private final TableExpressionRenderer tableExpressionRenderer;
     private final TableAliasCalculator tableAliasCalculator;
+    private final RenderingStrategy renderingStrategy;
+    private final AtomicInteger sequence;
 
     private JoinRenderer(Builder builder) {
         joinModel = Objects.requireNonNull(builder.joinModel);
         tableExpressionRenderer = Objects.requireNonNull(builder.tableExpressionRenderer);
         tableAliasCalculator = Objects.requireNonNull(builder.tableAliasCalculator);
+        renderingStrategy = Objects.requireNonNull(builder.renderingStrategy);
+        sequence = Objects.requireNonNull(builder.sequence);
     }
 
     public FragmentAndParameters render() {
@@ -50,26 +56,44 @@ public class JoinRenderer {
 
     private FragmentAndParameters renderJoinSpecification(JoinSpecification joinSpecification) {
         FragmentAndParameters renderedTable = joinSpecification.table().accept(tableExpressionRenderer);
+        FragmentAndParameters renderedJoin = renderConditions(joinSpecification);
 
         String fragment = joinSpecification.joinType().type()
                 + spaceBefore(renderedTable.fragment())
-                + spaceBefore(renderConditions(joinSpecification));
+                + spaceBefore(renderedJoin.fragment());
 
         return FragmentAndParameters.withFragment(fragment)
                 .withParameters(renderedTable.parameters())
+                .withParameters(renderedJoin.parameters())
                 .build();
     }
 
-    private String renderConditions(JoinSpecification joinSpecification) {
-        return joinSpecification.mapJoinCriteria(this::renderCriterion)
-                .collect(Collectors.joining(" ")); //$NON-NLS-1$
+    private FragmentAndParameters renderConditions(JoinSpecification joinSpecification) {
+        FragmentCollector fragmentCollector = joinSpecification.mapJoinCriteria(this::renderCriterion)
+                .collect(FragmentCollector.collect());
+
+        return FragmentAndParameters
+                .withFragment(fragmentCollector.fragments().collect(Collectors.joining(" "))) //$NON-NLS-1$
+                .withParameters(fragmentCollector.parameters())
+                .build();
     }
 
-    private String renderCriterion(JoinCriterion joinCriterion) {
-        return joinCriterion.connector()
-                + spaceBefore(applyTableAlias(joinCriterion.leftColumn()))
-                + spaceBefore(joinCriterion.operator())
-                + spaceBefore(applyTableAlias(joinCriterion.rightColumn()));
+    private <T> FragmentAndParameters renderCriterion(JoinCriterion<T> joinCriterion) {
+        String prefix = joinCriterion.connector()
+                + spaceBefore(applyTableAlias(joinCriterion.leftColumn()));
+
+        JoinConditionRenderer<T> joinConditionRenderer = new JoinConditionRenderer.Builder<T>()
+                .withRenderingStrategy(renderingStrategy)
+                .withSequence(sequence)
+                .withTableAliasCalculator(tableAliasCalculator)
+                .withLeftColumn(joinCriterion.leftColumn())
+                .build();
+
+        FragmentAndParameters suffix = joinCriterion.joinCondition().accept(joinConditionRenderer);
+
+        return FragmentAndParameters.withFragment(prefix + spaceBefore(suffix.fragment()))
+                .withParameters(suffix.parameters())
+                .build();
     }
 
     private String applyTableAlias(BasicColumn column) {
@@ -84,6 +108,8 @@ public class JoinRenderer {
         private JoinModel joinModel;
         private TableExpressionRenderer tableExpressionRenderer;
         private TableAliasCalculator tableAliasCalculator;
+        private RenderingStrategy renderingStrategy;
+        private AtomicInteger sequence;
 
         public Builder withJoinModel(JoinModel joinModel) {
             this.joinModel = joinModel;
@@ -97,6 +123,16 @@ public class JoinRenderer {
 
         public Builder withTableAliasCalculator(TableAliasCalculator tableAliasCalculator) {
             this.tableAliasCalculator = tableAliasCalculator;
+            return this;
+        }
+
+        public Builder withRenderingStrategy(RenderingStrategy renderingStrategy) {
+            this.renderingStrategy = renderingStrategy;
+            return this;
+        }
+
+        public Builder withSequence(AtomicInteger sequence) {
+            this.sequence = sequence;
             return this;
         }