+++ src/main/java/org/mybatis/dynamic/sql/select/render/QueryExpressionRenderer.java
@@ -171,6 +171,8 @@ public class QueryExpressionRenderer {
         return JoinRenderer.withJoinModel(joinModel)
                 .withTableExpressionRenderer(tableExpressionRenderer)
                 .withTableAliasCalculator(tableAliasCalculator)
+                .withRenderingStrategy(renderingStrategy)
+                .withSequence(sequence)
                 .build()
                 .render();
     }